backend/requirements.txt
# Backend - LexIntellectus API
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pypdf
python-docx


# AI & Vector Support
openai
google-generativeai
pgvector==0.2.5

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Redis
redis==5.0.1

# MinIO / S3
minio==7.2.3

# Utilities
email-validator==2.1.0.post1
httpx==0.27.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
